refactor: flattenTasks 중복 함수 제거
- shared/lib/taskAdapters에 flattenTasks 공통 유틸리티 추가
- GanttChartView, KanbanBoard, taskStore의 중복 코드 제거
- 3곳의 동일한 로직을 1곳으로 통합하여 유지보수성 향상
- JSDoc 주석으로 함수 목적 문서화

Resolves: https://github.com/DevOpsSociety/FlowPlan-front/pull/28#discussion_r2461688222
